AGENTIC AI WITH LLM/requirements.txt
langchain
langchain-groq
langchain-core
langgraph
fastapi
uvicorn
python-dotenv
